refactor: Split zero-day exploit tests into a module structure and update `BlockStorageService` constructor with a dependencies struct.
