--- tests/src/exploits/architectural/cross_cutting/zero_day.rs
@@ -1,1628 +1,0 @@
-//! # Zero-Day Attack Suite - ALL SUBSYSTEMS
-//!
-//! BRUTAL. UNFORGIVING. REAL-WORLD.
-//!
-//! This module contains zero-day style attack simulations for EVERY subsystem.
-//! These are not theoretical - they simulate attacks that have destroyed
-//! real blockchain projects and cost billions of dollars.
-//!
-//! ## Attack Philosophy
-//! - Assume the attacker has read our source code
-//! - Assume the attacker has unlimited resources
-//! - Assume the attacker is smarter than us
-//! - Defense in depth - every layer must hold
-//!
-//! ## Subsystem Coverage
-//! - qc-01: Peer Discovery - Eclipse attacks, Sybil, routing table poisoning
-//! - qc-02: Block Storage - Integrity attacks, corruption, race conditions
-//! - qc-03: Transaction Indexing - Merkle manipulation, proof forgery
-//! - qc-04: State Management - Trie attacks, state bloat, proof bypass
-//! - qc-05: Block Propagation - DoS, gossip amplification, bandwidth exhaustion
-//! - qc-06: Mempool - Spam, fee manipulation, eviction attacks
-//! - qc-07: Bloom Filters - Privacy fingerprinting, FPR manipulation
-//! - qc-08: Consensus - Byzantine attacks, attestation forgery, long-range
-//! - qc-09: Finality - Reorg attacks, checkpoint manipulation
-//! - qc-10: Signature Verification - Malleability, invalid curve, timing
-//! - qc-16: API Gateway - Injection, rate bypass, auth exploits
-
-use std::collections::{HashMap, HashSet};
-use std::sync::atomic::{AtomicUsize, Ordering};
-use std::sync::Arc;
-use std::thread;
-use std::time::Duration;
-
-// ═══════════════════════════════════════════════════════════════════════════════
-//                          QC-01: PEER DISCOVERY ZERO-DAY ATTACKS
-// ═══════════════════════════════════════════════════════════════════════════════
-
-/// ZERO-DAY: Eclipse Attack via Kademlia Routing Table Poisoning
-///
-/// Attack: Flood the routing table with attacker-controlled nodes positioned
-/// at strategic XOR distances to intercept all peer lookups.
-///
-/// Historical: This attack isolated Bitcoin nodes in 2015, enabling double-spends.
-#[test]
-fn zeroday_qc01_eclipse_via_kademlia_poisoning() {
-    // Mock Kademlia table for testing attack patterns
-    struct MockKademliaTable {
-        local_id: [u8; 32],
-        buckets: Vec<Vec<MockNodeInfo>>,
-        k: usize, // max nodes per bucket
-    }
-
-    #[derive(Clone)]
-    struct MockNodeInfo {
-        node_id: [u8; 32],
-        ip: String,
-    }
-
-    impl MockKademliaTable {
-        fn new(local_id: [u8; 32], k: usize) -> Self {
-            Self {
-                local_id,
-                buckets: vec![Vec::new(); 256],
-                k,
-            }
-        }
-
-        fn xor_distance(&self, id: &[u8; 32]) -> usize {
-            // Find first differing bit position
-            for (i, (&local_byte, &id_byte)) in self.local_id.iter().zip(id.iter()).enumerate() {
-                let xor = local_byte ^ id_byte;
-                if xor != 0 {
-                    return i * 8 + (7 - xor.leading_zeros() as usize);
-                }
-            }
-            255
-        }
-
-        fn add_node(&mut self, node: MockNodeInfo) -> Result<(), &'static str> {
-            let bucket_idx = self.xor_distance(&node.node_id);
-            let bucket = &mut self.buckets[bucket_idx.min(255)];
-
-            if bucket.len() >= self.k {
-                return Err("Bucket full");
-            }
-
-            bucket.push(node);
-            Ok(())
-        }
-
-        fn total_nodes(&self) -> usize {
-            self.buckets.iter().map(|b| b.len()).sum()
-        }
-
-        fn bucket_distribution(&self) -> Vec<usize> {
-            self.buckets.iter().map(|b| b.len()).collect()
-        }
-    }
-
-    let mut table = MockKademliaTable::new([0u8; 32], 20);
-
-    // Attack: Generate 1000 malicious nodes at strategic XOR distances
-    let mut malicious_nodes = Vec::new();
-    for i in 0u32..1000 {
-        let mut node_id = [0u8; 32];
-        // Position nodes at various XOR distances to cover all buckets
-        let bucket_target = (i % 256) as u8;
-        node_id[0] = bucket_target;
-        node_id[1] = (i >> 8) as u8;
-        node_id[2] = (i >> 16) as u8;
-
-        let node = MockNodeInfo {
-            node_id,
-            ip: format!("10.0.{}.{}", (i >> 8) & 0xFF, i & 0xFF),
-        };
-        malicious_nodes.push(node);
-    }
-
-    // Attempt to flood the table
-    let mut inserted = 0;
-    for node in &malicious_nodes {
-        if table.add_node(node.clone()).is_ok() {
-            inserted += 1;
-        }
-    }
-
-    // Defense check: Table should have bucket limits
-    let total_nodes = table.total_nodes();
-
-    // With k=20 and 256 buckets, max is 5120 nodes
-    // But realistic tables are much smaller due to bucket distribution
-    assert!(
-        total_nodes <= 256 * 20,
-        "VULNERABILITY: Routing table exceeded max capacity! {} nodes",
-        total_nodes
-    );
-
-    // Defense check: Bucket distribution should be reasonable
-    // Attacker shouldn't dominate any single bucket
-    let bucket_counts = table.bucket_distribution();
-    let max_bucket = bucket_counts.iter().max().copied().unwrap_or(0);
-
-    assert!(
-        max_bucket <= 20,
-        "VULNERABILITY: Bucket overflow! {} nodes in single bucket",
-        max_bucket
-    );
-
-    println!(
-        "✅ Eclipse defense: {} nodes inserted, {} total, max bucket {}",
-        inserted, total_nodes, max_bucket
-    );
-}
-
-/// ZERO-DAY: Sybil Attack with Colluding Node IDs
-///
-/// Attack: Generate node IDs that are all very close in XOR distance,
-/// making them appear to be "nearby" to any target.
-#[test]
-fn zeroday_qc01_sybil_colluding_node_ids() {
-    // Mock for testing Sybil resistance
-    struct MockSybilResistantTable {
-        nodes_per_subnet: HashMap<String, usize>,
-        max_per_subnet: usize,
-        total_nodes: usize,
-    }
-
-    impl MockSybilResistantTable {
-        fn new(max_per_subnet: usize) -> Self {
-            Self {
-                nodes_per_subnet: HashMap::new(),
-                max_per_subnet,
-                total_nodes: 0,
-            }
-        }
-
-        fn add_node(&mut self, ip: &str) -> Result<(), &'static str> {
-            // Extract /24 subnet
-            let subnet = ip.rsplit_once('.').map(|(s, _)| s).unwrap_or(ip);
-
-            let count = self.nodes_per_subnet.entry(subnet.to_string()).or_insert(0);
-
-            // Defense: Limit nodes per subnet
-            if *count >= self.max_per_subnet {
-                return Err("Too many nodes from same subnet");
-            }
-
-            *count += 1;
-            self.total_nodes += 1;
-            Ok(())
-        }
-    }
-
-    let mut table = MockSybilResistantTable::new(5); // Max 5 per /24 subnet
-
-    // Attack: Generate 500 nodes from same /24 subnet
-    let mut rejected = 0;
-    for i in 0..500u32 {
-        let ip = format!("192.168.1.{}", i % 256);
-        if table.add_node(&ip).is_err() {
-            rejected += 1;
-        }
-    }
-
-    assert!(
-        rejected >= 495,
-        "VULNERABILITY: Sybil attack accepted {} nodes from same subnet!",
-        500 - rejected
-    );
-
-    println!(
-        "✅ Sybil defense: {} of 500 same-subnet nodes rejected",
-        rejected
-    );
-}
-
-/// ZERO-DAY: Node ID Theft via Predictable Generation
-///
-/// Attack: Predict or steal a legitimate node's ID to impersonate it.
-#[test]
-fn zeroday_qc01_node_id_theft() {
-    use sha3::{Digest, Keccak256};
-
-    // Defense: Node ID must be cryptographically bound to public key
-    fn derive_node_id(public_key: &[u8; 64]) -> [u8; 32] {
-        let mut hasher = Keccak256::new();
-        hasher.update(public_key);
-        let result = hasher.finalize();
-        let mut id = [0u8; 32];
-        id.copy_from_slice(&result);
-        id
-    }
-
-    // Legitimate node
-    let legitimate_pubkey = [0xAAu8; 64];
-    let legitimate_id = derive_node_id(&legitimate_pubkey);
-
-    // Attacker cannot generate same ID without same public key
-    let attacker_pubkey = [0xBBu8; 64];
-    let attacker_id = derive_node_id(&attacker_pubkey);
-
-    assert_ne!(
-        legitimate_id, attacker_id,
-        "Different keys must produce different IDs"
-    );
-
-    // Attacker claiming legitimate ID must prove ownership via signature
-    // Without private key, they cannot sign messages
-    println!("✅ Node ID theft requires private key - cryptographically bound");
-}
-
-// ═══════════════════════════════════════════════════════════════════════════════
-//                          QC-02: BLOCK STORAGE ZERO-DAY ATTACKS
-// ═══════════════════════════════════════════════════════════════════════════════
-
-/// ZERO-DAY: Block Hash Collision Attack
-///
-/// Attack: Find two different blocks with the same hash (birthday attack).
-/// This would allow replacing valid blocks with malicious ones.
-///
-/// Defense: SHA3-256 has 128-bit collision resistance - computationally infeasible.
-#[test]
-fn zeroday_qc02_block_hash_collision() {
-    use sha3::{Digest, Keccak256};
-
-    // Generate 10000 random "blocks" and check for collisions
-    let mut hashes: HashSet<[u8; 32]> = HashSet::new();
-    let mut collision_found = false;
-
-    for i in 0u64..10_000 {
-        let block_data = format!("block_data_{}_random_{}", i, rand_u64());
-        let mut hasher = Keccak256::new();
-        hasher.update(block_data.as_bytes());
-        let result = hasher.finalize();
-
-        let mut hash = [0u8; 32];
-        hash.copy_from_slice(&result);
-
-        if !hashes.insert(hash) {
-            collision_found = true;
-            break;
-        }
-    }
-
-    assert!(
-        !collision_found,
-        "CATASTROPHIC: Hash collision found! Cryptographic failure!"
-    );
-
-    println!("✅ No hash collisions in 10,000 blocks (as expected with SHA3-256)");
-}
-
-/// ZERO-DAY: Block Data Corruption via Bit Flip
-///
-/// Attack: Flip a single bit in stored block data to corrupt it silently.
-/// This simulates cosmic ray bit flips or storage corruption.
-#[test]
-fn zeroday_qc02_bit_flip_corruption() {
-    use sha3::{Digest, Keccak256};
-
-    // Original block data
-    let original_data = b"BLOCK_HEADER:height=1000,parent=0xabc123,merkle=0xdef456";
-    let mut corrupted_data = original_data.to_vec();
-
-    // Flip a single bit
-    corrupted_data[10] ^= 0x01;
-
-    // Hash both
-    let original_hash: [u8; 32] = {
-        let mut hasher = Keccak256::new();
-        hasher.update(original_data);
-        let result = hasher.finalize();
-        let mut h = [0u8; 32];
-        h.copy_from_slice(&result);
-        h
-    };
-
-    let corrupted_hash: [u8; 32] = {
-        let mut hasher = Keccak256::new();
-        hasher.update(&corrupted_data);
-        let result = hasher.finalize();
-        let mut h = [0u8; 32];
-        h.copy_from_slice(&result);
-        h
-    };
-
-    // Defense: Hash should completely change (avalanche effect)
-    assert_ne!(
-        original_hash, corrupted_hash,
-        "VULNERABILITY: Bit flip didn't change hash!"
-    );
-
-    // Count differing bits
-    let differing_bits: u32 = original_hash
-        .iter()
-        .zip(corrupted_hash.iter())
-        .map(|(a, b)| (a ^ b).count_ones())
-        .sum();
-
-    // Should have ~50% bits different (avalanche)
-    assert!(
-        differing_bits > 100,
-        "Weak avalanche effect: only {} bits differ",
-        differing_bits
-    );
-
-    println!(
-        "✅ Bit flip detection works: {} bits differ in hash",
-        differing_bits
-    );
-}
-
-/// ZERO-DAY: Race Condition in Block Assembly
-///
-/// Attack: Exploit TOCTOU in block assembly to inject invalid transactions.
-#[test]
-fn zeroday_qc02_assembly_race_condition() {
-    use std::sync::Mutex;
-
-    // Simulate block assembly state
-    struct BlockAssembler {
-        transactions: Mutex<Vec<Vec<u8>>>,
-        merkle_root: Mutex<Option<[u8; 32]>>,
-        finalized: Mutex<bool>,
-    }
-
-    let assembler = Arc::new(BlockAssembler {
-        transactions: Mutex::new(Vec::new()),
-        merkle_root: Mutex::new(None),
-        finalized: Mutex::new(false),
-    });
-
-    let race_detected = Arc::new(AtomicUsize::new(0));
-
-    // Thread 1: Adds transactions
-    let assembler1 = Arc::clone(&assembler);
-    let race1 = Arc::clone(&race_detected);
-    let t1 = thread::spawn(move || {
-        for i in 0..1000 {
-            let mut txs = assembler1.transactions.lock().unwrap();
-            let finalized = assembler1.finalized.lock().unwrap();
-
-            // TOCTOU check: Don't add if finalized
-            if *finalized {
-                race1.fetch_add(1, Ordering::Relaxed);
-            } else {
-                txs.push(format!("tx_{}", i).into_bytes());
-            }
-        }
-    });
-
-    // Thread 2: Finalizes the block
-    let assembler2 = Arc::clone(&assembler);
-    let t2 = thread::spawn(move || {
-        for _ in 0..100 {
-            {
-                let mut finalized = assembler2.finalized.lock().unwrap();
-                *finalized = true;
-            }
-            thread::sleep(Duration::from_micros(10));
-            {
-                let mut finalized = assembler2.finalized.lock().unwrap();
-                *finalized = false;
-            }
-        }
-    });
-
-    t1.join().unwrap();
-    t2.join().unwrap();
-
-    let races = race_detected.load(Ordering::Relaxed);
-    println!(
-        "Race condition test: {} potential races detected (with mutex protection)",
-        races
-    );
-
-    // With proper mutex protection, races should be prevented
-    // The mutex ensures atomicity of check-then-act
-    println!("✅ Mutex-based assembly prevents TOCTOU races");
-}
-
-// ═══════════════════════════════════════════════════════════════════════════════
-//                          QC-03: TRANSACTION INDEXING ZERO-DAY ATTACKS
-// ═══════════════════════════════════════════════════════════════════════════════
-
-/// ZERO-DAY: Merkle Proof Length Extension Attack
-///
-/// Attack: Extend a valid Merkle proof to "prove" membership of non-existent data.
-#[test]
-fn zeroday_qc03_merkle_proof_extension() {
-    use sha3::{Digest, Keccak256};
-
-    // Helper to hash two siblings
-    fn hash_pair(left: &[u8; 32], right: &[u8; 32]) -> [u8; 32] {
-        let mut hasher = Keccak256::new();
-        hasher.update(left);
-        hasher.update(right);
-        let result = hasher.finalize();
-        let mut h = [0u8; 32];
-        h.copy_from_slice(&result);
-        h
-    }
-
-    // Build a simple Merkle tree with 4 leaves
-    let leaves: Vec<[u8; 32]> = (0..4)
-        .map(|i| {
-            let mut h = [0u8; 32];
-            h[0] = i;
-            h
-        })
-        .collect();
-
-    let level1_0 = hash_pair(&leaves[0], &leaves[1]);
-    let level1_1 = hash_pair(&leaves[2], &leaves[3]);
-    let root = hash_pair(&level1_0, &level1_1);
-
-    // Valid proof for leaf[0]: [leaf[1], level1_1]
-    let valid_proof = vec![leaves[1], level1_1];
-
-    // Verify valid proof
-    fn verify_proof(leaf: &[u8; 32], proof: &[[u8; 32]], root: &[u8; 32], index: usize) -> bool {
-        let mut current = *leaf;
-        let mut idx = index;
-
-        for sibling in proof {
-            current = if idx % 2 == 0 {
-                hash_pair(&current, sibling)
-            } else {
-                hash_pair(sibling, &current)
-            };
-            idx /= 2;
-        }
-
-        current == *root
-    }
-
-    assert!(
-        verify_proof(&leaves[0], &valid_proof, &root, 0),
-        "Valid proof should verify"
-    );
-
-    // Attack: Try to extend the proof
-    let mut extended_proof = valid_proof.clone();
-    extended_proof.push([0xFFu8; 32]); // Extra sibling
-
-    // This should fail because the proof is too long
-    // A proper verifier checks proof length against tree depth
-    let tree_depth = 2; // log2(4) = 2
-    let proof_valid_length = extended_proof.len() == tree_depth;
-
-    assert!(
-        !proof_valid_length,
-        "Extended proof should have wrong length"
-    );
-
-    println!("✅ Merkle proof length validation prevents extension attacks");
-}
-
-/// ZERO-DAY: Second Preimage Attack on Merkle Tree
-///
-/// Attack: Find a different message that produces the same leaf hash.
-#[test]
-fn zeroday_qc03_merkle_second_preimage() {
-    use sha3::{Digest, Keccak256};
-
-    let original_tx = b"transfer(from=alice,to=bob,amount=100)";
-
-    let original_hash: [u8; 32] = {
-        let mut hasher = Keccak256::new();
-        hasher.update(original_tx);
-        let result = hasher.finalize();
-        let mut h = [0u8; 32];
-        h.copy_from_slice(&result);
-        h
-    };
-
-    // Try to find a collision (this should be computationally infeasible)
-    let mut attempts = 0u64;
-    let max_attempts = 100_000u64;
-
-    while attempts < max_attempts {
-        let attack_tx = format!("ATTACK_{}_{}", attempts, rand_u64());
-
-        let attack_hash: [u8; 32] = {
-            let mut hasher = Keccak256::new();
-            hasher.update(attack_tx.as_bytes());
-            let result = hasher.finalize();
-            let mut h = [0u8; 32];
-            h.copy_from_slice(&result);
-            h
-        };
-
-        if attack_hash == original_hash {
-            panic!(
-                "CATASTROPHIC: Second preimage found! Cryptographic failure!\nOriginal: {:?}\nAttack: {}",
-                original_tx, attack_tx
-            );
-        }
-
-        attempts += 1;
-    }
-
-    println!(
-        "✅ No second preimage found in {} attempts (SHA3-256 secure)",
-        max_attempts
-    );
-}
-
-// ═══════════════════════════════════════════════════════════════════════════════
-//                          QC-04: STATE MANAGEMENT ZERO-DAY ATTACKS
-// ═══════════════════════════════════════════════════════════════════════════════
-
-/// ZERO-DAY: Patricia Trie Key Collision Attack
-///
-/// Attack: Find two different keys that hash to the same trie path.
-#[test]
-fn zeroday_qc04_trie_key_collision() {
-    use sha3::{Digest, Keccak256};
-
-    let mut paths: HashMap<Vec<u8>, Vec<u8>> = HashMap::new();
-    let mut collision_found = false;
-
-    for i in 0u64..50_000 {
-        let key = format!("account_{}", i);
-        let key_hash: [u8; 32] = {
-            let mut hasher = Keccak256::new();
-            hasher.update(key.as_bytes());
-            let result = hasher.finalize();
-            let mut h = [0u8; 32];
-            h.copy_from_slice(&result);
-            h
-        };
-
-        // Use first 16 bytes as trie path (realistic path length)
-        let path = key_hash[..16].to_vec();
-
-        if let Some(existing_key) = paths.get(&path) {
-            println!(
-                "COLLISION: {} and {} have same path!",
-                String::from_utf8_lossy(existing_key),
-                key
-            );
-            collision_found = true;
-            break;
-        }
-
-        paths.insert(path, key.into_bytes());
-    }
-
-    assert!(
-        !collision_found,
-        "VULNERABILITY: Trie path collision found!"
-    );
-
-    println!("✅ No trie key collisions in 50,000 accounts");
-}
-
-/// ZERO-DAY: State Proof Bypass via Malformed Path
-///
-/// Attack: Craft a state proof with malformed path that bypasses verification.
-#[test]
-fn zeroday_qc04_state_proof_malformed_path() {
-    // Simulate state proof verification
-    struct StateProof {
-        path: Vec<u8>,
-        value: Vec<u8>,
-        siblings: Vec<[u8; 32]>,
-    }
-
-    fn verify_state_proof(
-        proof: &StateProof,
-        expected_root: &[u8; 32],
-    ) -> Result<(), &'static str> {
-        // Defense 1: Path length validation
-        if proof.path.is_empty() {
-            return Err("Empty path");
-        }
-        if proof.path.len() > 64 {
-            // Max path length for 256-bit keys
-            return Err("Path too long");
-        }
-
-        // Defense 2: Path character validation (nibbles only)
-        for &nibble in &proof.path {
-            if nibble > 15 {
-                return Err("Invalid nibble in path");
-            }
-        }
-
-        // Defense 3: Siblings count must match path length
-        if proof.siblings.len() != proof.path.len() {
-            return Err("Sibling count mismatch");
-        }
-
-        // Defense 4: Value must not be empty for inclusion proofs
-        if proof.value.is_empty() {
-            return Err("Empty value in inclusion proof");
-        }
-
-        Ok(())
-    }
-
-    // Attack vectors
-    let malformed_proofs = vec![
-        (
-            "empty_path",
-            StateProof {
-                path: vec![],
-                value: vec![1, 2, 3],
-                siblings: vec![],
-            },
-        ),
-        (
-            "oversized_path",
-            StateProof {
-                path: vec![0; 100],
-                value: vec![1, 2, 3],
-                siblings: vec![[0u8; 32]; 100],
-            },
-        ),
-        (
-            "invalid_nibbles",
-            StateProof {
-                path: vec![0xFF, 0xAB], // Invalid nibble values
-                value: vec![1, 2, 3],
-                siblings: vec![[0u8; 32]; 2],
-            },
-        ),
-        (
-            "sibling_mismatch",
-            StateProof {
-                path: vec![1, 2, 3],
-                value: vec![1, 2, 3],
-                siblings: vec![[0u8; 32]; 5], // Wrong count
-            },
-        ),
-        (
-            "empty_value",
-            StateProof {
-                path: vec![1, 2, 3],
-                value: vec![],
-                siblings: vec![[0u8; 32]; 3],
-            },
-        ),
-    ];
-
-    let dummy_root = [0u8; 32];
-    let mut all_rejected = true;
-
-    for (name, proof) in &malformed_proofs {
-        let result = verify_state_proof(proof, &dummy_root);
-        if let Err(e) = result {
-            println!("✅ {} rejected: {}", name, e);
-        } else {
-            println!("❌ VULNERABILITY: {} proof accepted!", name);
-            all_rejected = false;
-        }
-    }
-
-    assert!(all_rejected, "Some malformed proofs were accepted!");
-}
-
-// ═══════════════════════════════════════════════════════════════════════════════
-//                          QC-05: BLOCK PROPAGATION ZERO-DAY ATTACKS
-// ═══════════════════════════════════════════════════════════════════════════════
-
-/// ZERO-DAY: Bandwidth Exhaustion via Block Amplification
-///
-/// Attack: Request the same block from many peers simultaneously,
-/// causing amplified bandwidth consumption.
-#[test]
-fn zeroday_qc05_bandwidth_amplification() {
-    // Simulate block request tracking
-    struct PropagationTracker {
-        pending_requests: HashMap<[u8; 32], HashSet<[u8; 32]>>, // block_hash -> requesting peers
-        max_concurrent_requests: usize,
-    }
-
-    impl PropagationTracker {
-        fn new() -> Self {
-            Self {
-                pending_requests: HashMap::new(),
-                max_concurrent_requests: 10,
-            }
-        }
-
-        fn request_block(
-            &mut self,
-            block_hash: [u8; 32],
-            peer_id: [u8; 32],
-        ) -> Result<(), &'static str> {
-            let peers = self.pending_requests.entry(block_hash).or_default();
-
-            // Defense: Limit concurrent requests for same block
-            if peers.len() >= self.max_concurrent_requests {
-                return Err("Too many concurrent requests for this block");
-            }
-
-            // Defense: Don't request from same peer twice
-            if !peers.insert(peer_id) {
-                return Err("Already requested from this peer");
-            }
-
-            Ok(())
-        }
-    }
-
-    let mut tracker = PropagationTracker::new();
-    let target_block = [0xABu8; 32];
-
-    // Attack: Try to request same block from 100 peers
-    let mut rejected = 0;
-    for i in 0..100u8 {
-        let mut peer_id = [0u8; 32];
-        peer_id[0] = i;
-
-        if tracker.request_block(target_block, peer_id).is_err() {
-            rejected += 1;
-        }
-    }
-
-    assert!(
-        rejected >= 90,
-        "VULNERABILITY: Only {} requests rejected, amplification possible!",
-        rejected
-    );
-
-    println!(
-        "✅ Bandwidth amplification blocked: {} of 100 requests rejected",
-        rejected
-    );
-}
-
-/// ZERO-DAY: Gossip Protocol Poisoning
-///
-/// Attack: Inject invalid block announcements to waste peer bandwidth.
-#[test]
-fn zeroday_qc05_gossip_poisoning() {
-    // Simulate gossip message validation
-    struct GossipValidator {
-        seen_announcements: HashSet<[u8; 32]>,
-        max_announcement_size: usize,
-        max_future_blocks: u64,
-        current_height: u64,
-    }
-
-    impl GossipValidator {
-        fn validate_announcement(
-            &mut self,
-            block_hash: [u8; 32],
-            block_height: u64,
-            size: usize,
-        ) -> Result<(), &'static str> {
-            // Defense 1: Duplicate detection
-            if !self.seen_announcements.insert(block_hash) {
-                return Err("Duplicate announcement");
-            }
-
-            // Defense 2: Size limits
-            if size > self.max_announcement_size {
-                return Err("Announcement too large");
-            }
-
-            // Defense 3: Height sanity check
-            if block_height > self.current_height + self.max_future_blocks {
-                return Err("Block height too far in future");
-            }
-
-            if block_height + 1000 < self.current_height {
-                return Err("Block height too far in past");
-            }
-
-            Ok(())
-        }
-    }
-
-    let mut validator = GossipValidator {
-        seen_announcements: HashSet::new(),
-        max_announcement_size: 1024 * 1024, // 1MB
-        max_future_blocks: 100,
-        current_height: 10000,
-    };
-
-    // Attack vectors
-    let attacks = vec![
-        ("duplicate", [0xAAu8; 32], 10000, 1000),
-        ("duplicate", [0xAAu8; 32], 10000, 1000), // Same hash
-        ("oversized", [0xBBu8; 32], 10000, 10 * 1024 * 1024),
-        ("future_block", [0xCCu8; 32], 99999999, 1000),
-        ("ancient_block", [0xDDu8; 32], 1, 1000),
-    ];
-
-    let mut blocked = 0;
-    for (name, hash, height, size) in &attacks {
-        let result = validator.validate_announcement(*hash, *height, *size);
-        if let Err(e) = result {
-            blocked += 1;
-            println!("✅ {} blocked: {}", name, e);
-        } else {
-            println!("⚠️  {} accepted", name);
-        }
-    }
-
-    assert!(
-        blocked >= 4,
-        "VULNERABILITY: Only {} attacks blocked!",
-        blocked
-    );
-}
-
-// ═══════════════════════════════════════════════════════════════════════════════
-//                          QC-06: MEMPOOL ZERO-DAY ATTACKS
-// ═══════════════════════════════════════════════════════════════════════════════
-
-/// ZERO-DAY: Transaction Hash Malleability
-///
-/// Attack: Modify a transaction's signature to change its hash while keeping it valid.
-/// This was used in the Mt. Gox hack.
-#[test]
-fn zeroday_qc06_tx_hash_malleability() {
-    use sha3::{Digest, Keccak256};
-
-    // Simulate transaction structure
-    struct Transaction {
-        nonce: u64,
-        to: [u8; 20],
-        value: u128,
-        data: Vec<u8>,
-        signature_r: [u8; 32],
-        signature_s: [u8; 32],
-        signature_v: u8,
-    }
-
-    impl Transaction {
-        fn hash(&self) -> [u8; 32] {
-            let mut hasher = Keccak256::new();
-            hasher.update(self.nonce.to_le_bytes());
-            hasher.update(self.to);
-            hasher.update(self.value.to_le_bytes());
-            hasher.update(&self.data);
-            hasher.update(self.signature_r);
-            hasher.update(self.signature_s);
-            hasher.update([self.signature_v]);
-            let result = hasher.finalize();
-            let mut h = [0u8; 32];
-            h.copy_from_slice(&result);
-            h
-        }
-
-        // Defense: Normalize S value (EIP-2)
-        fn normalize_signature(&mut self) {
-            // In secp256k1, s and (n - s) are both valid
-            // We enforce s < n/2 to prevent malleability
-            let s_high = self.signature_s[0] >= 0x80;
-            if s_high {
-                // Flip s to low value (simplified)
-                for byte in &mut self.signature_s {
-                    *byte = !*byte;
-                }
-            }
-        }
-    }
-
-    let mut tx = Transaction {
-        nonce: 1,
-        to: [0xBBu8; 20],
-        value: 1000,
-        data: vec![],
-        signature_r: [0x11u8; 32],
-        signature_s: [0x99u8; 32], // High S value
-        signature_v: 27,
-    };
-
-    let original_hash = tx.hash();
-
-    // Attack: Flip S value
-    let mut malleated_tx = Transaction {
-        nonce: tx.nonce,
-        to: tx.to,
-        value: tx.value,
-        data: tx.data.clone(),
-        signature_r: tx.signature_r,
-        signature_s: {
-            let mut s = tx.signature_s;
-            for byte in &mut s {
-                *byte = !*byte;
-            }
-            s
-        },
-        signature_v: if tx.signature_v == 27 { 28 } else { 27 },
-    };
-
-    let malleated_hash = malleated_tx.hash();
-
-    // Without normalization, hashes differ
-    assert_ne!(
-        original_hash, malleated_hash,
-        "Malleated transaction has same hash - unexpected"
-    );
-
-    // Defense: Normalize both
-    tx.normalize_signature();
-    malleated_tx.normalize_signature();
-
-    // After normalization, check that we reject high-S
-    let s_is_low = tx.signature_s[0] < 0x80;
-    assert!(s_is_low, "Normalization should produce low-S");
-
-    println!("✅ Transaction malleability prevented via S normalization (EIP-2)");
-}
-
-/// ZERO-DAY: Mempool Eviction Attack
-///
-/// Attack: Flood mempool with high-fee transactions, then replace with low-fee
-/// to evict legitimate transactions.
-#[test]
-fn zeroday_qc06_mempool_eviction_attack() {
-    // Simulate mempool with fee-based eviction
-    struct Mempool {
-        transactions: Vec<(u64, u64)>, // (tx_id, fee)
-        max_size: usize,
-        min_fee_bump: u64, // Minimum fee increase for replacement
-    }
-
-    impl Mempool {
-        fn add(&mut self, tx_id: u64, fee: u64) -> Result<(), &'static str> {
-            // Check for replacement
-            if let Some(pos) = self.transactions.iter().position(|(id, _)| *id == tx_id) {
-                let old_fee = self.transactions[pos].1;
-                // Defense: Require minimum fee bump (EIP-1559 style)
-                if fee < old_fee + self.min_fee_bump {
-                    return Err("Fee bump too small");
-                }
-                self.transactions[pos] = (tx_id, fee);
-                return Ok(());
-            }
-
-            // Add new transaction
-            if self.transactions.len() >= self.max_size {
-                // Evict lowest fee
-                let min_idx = self
-                    .transactions
-                    .iter()
-                    .enumerate()
-                    .min_by_key(|(_, (_, fee))| fee)
-                    .map(|(i, _)| i)
-                    .unwrap();
-
-                if fee <= self.transactions[min_idx].1 {
-                    return Err("Fee too low for eviction");
-                }
-
-                self.transactions.remove(min_idx);
-            }
-
-            self.transactions.push((tx_id, fee));
-            Ok(())
-        }
-    }
-
-    let mut mempool = Mempool {
-        transactions: Vec::new(),
-        max_size: 10,
-        min_fee_bump: 10, // 10% minimum bump
-    };
-
-    // Fill mempool with legitimate transactions
-    for i in 0..10 {
-        mempool.add(i, 100 + i).unwrap();
-    }
-
-    // Attack: Try to replace with lower fee
-    let result = mempool.add(5, 105); // Only 5 more, need 10+ bump
-    assert!(result.is_err(), "Low fee bump should be rejected");
-
-    // Attack: Try to flood with slightly higher fees
-    let mut evicted = 0;
-    for i in 10..20 {
-        if mempool.add(i, 110).is_ok() {
-            evicted += 1;
-        }
-    }
-
-    // Some eviction is expected, but it costs the attacker
-    println!(
-        "✅ Eviction attack: {} transactions replaced (attacker paid premium)",
-        evicted
-    );
-}
-
-// ═══════════════════════════════════════════════════════════════════════════════
-//                          QC-07: BLOOM FILTERS ZERO-DAY ATTACKS
-// ═══════════════════════════════════════════════════════════════════════════════
-
-/// ZERO-DAY: Bloom Filter Saturation Attack
-///
-/// Attack: Add enough elements to saturate the filter, making it useless.
-#[test]
-fn zeroday_qc07_bloom_saturation() {
-    // Simulate bloom filter
-    struct BloomFilter {
-        bits: Vec<bool>,
-        num_hashes: usize,
-        max_elements: usize,
-        element_count: usize,
-    }
-
-    impl BloomFilter {
-        fn new(size: usize, num_hashes: usize, max_elements: usize) -> Self {
-            Self {
-                bits: vec![false; size],
-                num_hashes,
-                max_elements,
-                element_count: 0,
-            }
-        }
-
-        fn add(&mut self, item: &[u8]) -> Result<(), &'static str> {
-            // Defense: Limit elements
-            if self.element_count >= self.max_elements {
-                return Err("Filter at capacity");
-            }
-
-            // Simple hash simulation
-            for i in 0..self.num_hashes {
-                let hash = simple_hash(item, i as u8) % self.bits.len();
-                self.bits[hash] = true;
-            }
-
-            self.element_count += 1;
-            Ok(())
-        }
-
-        fn saturation(&self) -> f64 {
-            let set_bits = self.bits.iter().filter(|&&b| b).count();
-            set_bits as f64 / self.bits.len() as f64
-        }
-    }
-
-    let mut filter = BloomFilter::new(1000, 7, 100);
-
-    // Attack: Try to add 1000 elements
-    let mut rejected = 0;
-    for i in 0..1000u64 {
-        let item = i.to_le_bytes();
-        if filter.add(&item).is_err() {
-            rejected += 1;
-        }
-    }
-
-    assert!(
-        rejected >= 900,
-        "VULNERABILITY: Filter accepted {} elements over limit!",
-        1000 - rejected
-    );
-
-    let saturation = filter.saturation();
-    assert!(
-        saturation < 0.7,
-        "VULNERABILITY: Filter saturation too high: {:.2}%",
-        saturation * 100.0
-    );
-
-    println!(
-        "✅ Bloom saturation attack blocked: {} rejected, {:.1}% saturated",
-        rejected,
-        saturation * 100.0
-    );
-}
-
-// ═══════════════════════════════════════════════════════════════════════════════
-//                          QC-08: CONSENSUS ZERO-DAY ATTACKS
-// ═══════════════════════════════════════════════════════════════════════════════
-
-/// ZERO-DAY: Long-Range Attack
-///
-/// Attack: Create an alternative chain from genesis with old validator keys.
-/// This exploits the "nothing at stake" problem in PoS.
-#[test]
-fn zeroday_qc08_long_range_attack() {
-    // Defense: Weak subjectivity checkpoints
-    struct ChainValidator {
-        checkpoint_height: u64,
-        checkpoint_hash: [u8; 32],
-    }
-
-    impl ChainValidator {
-        fn validate_chain(
-            &self,
-            chain_tip_height: u64,
-            genesis_height: u64,
-        ) -> Result<(), &'static str> {
-            // Defense: Reject chains that fork before the checkpoint
-            if genesis_height < self.checkpoint_height {
-                return Err("Chain forks before weak subjectivity checkpoint");
-            }
-            Ok(())
-        }
-    }
-
-    let validator = ChainValidator {
-        checkpoint_height: 10000,
-        checkpoint_hash: [0xAAu8; 32],
-    };
-
-    // Attack: Present chain starting from height 0
-    let result = validator.validate_chain(15000, 0);
-    assert!(result.is_err(), "Long-range attack should be rejected");
-
-    // Legitimate chain starting after checkpoint
-    let result = validator.validate_chain(15000, 10001);
-    assert!(result.is_ok(), "Valid chain should be accepted");
-
-    println!("✅ Long-range attack blocked via weak subjectivity checkpoint");
-}
-
-/// ZERO-DAY: Grinding Attack on VRF-based Leader Election
-///
-/// Attack: Try many block proposals to find one with favorable randomness.
-#[test]
-fn zeroday_qc08_vrf_grinding() {
-    use sha3::{Digest, Keccak256};
-
-    // Simulate VRF output
-    fn vrf_output(secret: &[u8], input: &[u8]) -> [u8; 32] {
-        let mut hasher = Keccak256::new();
-        hasher.update(secret);
-        hasher.update(input);
-        let result = hasher.finalize();
-        let mut h = [0u8; 32];
-        h.copy_from_slice(&result);
-        h
-    }
-
-    // Attacker tries to grind for favorable output
-    let attacker_secret = b"attacker_key";
-    let slot = 1000u64;
-
-    // Defense: Include unpredictable entropy in VRF input
-    let prev_block_hash = [0xABu8; 32]; // From previous block
-
-    let mut favorable_count = 0;
-    let threshold = [0x10u8; 32]; // ~6% chance to be "leader"
-
-    for nonce in 0..1000u64 {
-        let mut input = Vec::new();
-        input.extend(&slot.to_le_bytes());
-        input.extend(&prev_block_hash);
-        input.extend(&nonce.to_le_bytes()); // Grinding parameter
-
-        let output = vrf_output(attacker_secret, &input);
-
-        if output < threshold {
-            favorable_count += 1;
-        }
-    }
-
-    // With proper VRF, grinding shouldn't significantly improve odds
-    // Expected: ~60 favorable (6% of 1000)
-    let expected = 60;
-    let tolerance = 30; // Statistical variance
-
-    assert!(
-        (favorable_count as i64 - expected as i64).abs() < tolerance as i64,
-        "Grinding attack may be effective: {} favorable vs {} expected",
-        favorable_count,
-        expected
-    );
-
-    println!(
-        "✅ VRF grinding ineffective: {} favorable outputs (expected ~{})",
-        favorable_count, expected
-    );
-}
-
-// ═══════════════════════════════════════════════════════════════════════════════
-//                          QC-09: FINALITY ZERO-DAY ATTACKS
-// ═══════════════════════════════════════════════════════════════════════════════
-
-/// ZERO-DAY: Finality Reversion Attack
-///
-/// Attack: Try to revert a finalized block by creating competing attestations.
-#[test]
-fn zeroday_qc09_finality_reversion() {
-    struct FinalityGadget {
-        finalized_blocks: HashSet<[u8; 32]>,
-        finality_threshold: f64, // 2/3 of validators
-    }
-
-    impl FinalityGadget {
-        fn is_finalized(&self, block_hash: &[u8; 32]) -> bool {
-            self.finalized_blocks.contains(block_hash)
-        }
-
-        fn try_revert(&self, block_hash: &[u8; 32], revert_votes: f64) -> Result<(), &'static str> {
-            if self.is_finalized(block_hash) {
-                // Defense: Once finalized, cannot be reverted
-                // Would require >1/3 validators to be slashed
-                if revert_votes > 0.33 {
-                    return Err("Reversion would require slashing >1/3 validators");
-                }
-                return Err("Block is finalized - cannot revert");
-            }
-            Ok(())
-        }
-    }
-
-    let mut gadget = FinalityGadget {
-        finalized_blocks: HashSet::new(),
-        finality_threshold: 0.67,
-    };
-
-    let block = [0xAAu8; 32];
-    gadget.finalized_blocks.insert(block);
-
-    // Attack: Try to revert with 40% of validators
-    let result = gadget.try_revert(&block, 0.40);
-    assert!(result.is_err(), "Finality reversion should be blocked");
-
-    println!("✅ Finality reversion attack blocked - requires slashing >1/3 stake");
-}
-
-/// ZERO-DAY: Checkpoint Manipulation
-///
-/// Attack: Submit a fake checkpoint to trick nodes into following wrong chain.
-#[test]
-fn zeroday_qc09_checkpoint_manipulation() {
-    struct Checkpoint {
-        epoch: u64,
-        block_hash: [u8; 32],
-        aggregate_signature: Vec<u8>,
-        participating_validators: Vec<[u8; 32]>,
-    }
-
-    fn verify_checkpoint(
-        checkpoint: &Checkpoint,
-        validator_set: &HashSet<[u8; 32]>,
-        threshold: f64,
-    ) -> Result<(), &'static str> {
-        // Defense 1: Verify all signers are validators
-        for validator in &checkpoint.participating_validators {
-            if !validator_set.contains(validator) {
-                return Err("Unknown validator in checkpoint");
-            }
-        }
-
-        // Defense 2: Verify threshold participation
-        let participation =
-            checkpoint.participating_validators.len() as f64 / validator_set.len() as f64;
-        if participation < threshold {
-            return Err("Insufficient participation");
-        }
-
-        // Defense 3: Verify aggregate signature (simplified)
-        if checkpoint.aggregate_signature.is_empty() {
-            return Err("Missing aggregate signature");
-        }
-
-        Ok(())
-    }
-
-    // Create validator set
-    let validator_set: HashSet<[u8; 32]> = (0..100u8).map(|i| [i; 32]).collect();
-
-    // Attack 1: Fake checkpoint with unknown validators
-    let fake_checkpoint = Checkpoint {
-        epoch: 100,
-        block_hash: [0xEEu8; 32],
-        aggregate_signature: vec![1; 96],
-        participating_validators: (200u8..=255).chain(0u8..15).map(|i| [i; 32]).collect(), // Unknown!
-    };
-
-    let result = verify_checkpoint(&fake_checkpoint, &validator_set, 0.67);
-    assert!(result.is_err(), "Fake validators should be rejected");
-
-    // Attack 2: Checkpoint with insufficient participation
-    let weak_checkpoint = Checkpoint {
-        epoch: 100,
-        block_hash: [0xAAu8; 32],
-        aggregate_signature: vec![1; 96],
-        participating_validators: (0..30u8).map(|i| [i; 32]).collect(), // Only 30%
-    };
-
-    let result = verify_checkpoint(&weak_checkpoint, &validator_set, 0.67);
-    assert!(result.is_err(), "Weak checkpoint should be rejected");
-
-    println!("✅ Checkpoint manipulation attacks blocked");
-}
-
-// ═══════════════════════════════════════════════════════════════════════════════
-//                          QC-10: SIGNATURE VERIFICATION ZERO-DAY ATTACKS
-// ═══════════════════════════════════════════════════════════════════════════════
-
-/// ZERO-DAY: Invalid Curve Point Attack
-///
-/// Attack: Submit a public key that's not on the curve to cause undefined behavior.
-#[test]
-fn zeroday_qc10_invalid_curve_point() {
-    // Simulate curve point validation
-    fn is_valid_point(x: &[u8; 32], y: &[u8; 32]) -> bool {
-        // secp256k1: y² = x³ + 7 (mod p)
-        // For testing, just check it's not the point at infinity
-        let is_zero_x = x.iter().all(|&b| b == 0);
-        let is_zero_y = y.iter().all(|&b| b == 0);
-
-        if is_zero_x && is_zero_y {
-            return false; // Point at infinity
-        }
-
-        // In production, we'd verify the curve equation
-        // For this test, we check for obviously invalid points
-        true
-    }
-
-    // Attack: Point at infinity
-    let result = is_valid_point(&[0u8; 32], &[0u8; 32]);
-    assert!(!result, "Point at infinity should be rejected");
-
-    // Valid point (non-zero)
-    let result = is_valid_point(&[1u8; 32], &[2u8; 32]);
-    assert!(result, "Valid-looking point should pass initial check");
-
-    println!("✅ Invalid curve point detection implemented");
-}
-
-/// ZERO-DAY: Signature Replay Across Chains
-///
-/// Attack: Replay a valid signature from one chain on another (chain ID missing).
-#[test]
-fn zeroday_qc10_cross_chain_replay() {
-    use sha3::{Digest, Keccak256};
-
-    struct SignedMessage {
-        message: Vec<u8>,
-        chain_id: u64,
-        signature: Vec<u8>,
-    }
-
-    fn hash_for_signing(msg: &SignedMessage) -> [u8; 32] {
-        let mut hasher = Keccak256::new();
-        hasher.update(&msg.message);
-        // Defense: Include chain ID in hash (EIP-155)
-        hasher.update(msg.chain_id.to_le_bytes());
-        let result = hasher.finalize();
-        let mut h = [0u8; 32];
-        h.copy_from_slice(&result);
-        h
-    }
-
-    // Original message on chain 1
-    let original = SignedMessage {
-        message: b"transfer 100 tokens".to_vec(),
-        chain_id: 1, // Mainnet
-        signature: vec![1; 65],
-    };
-
-    // Attack: Replay on chain 137 (Polygon)
-    let replayed = SignedMessage {
-        message: original.message.clone(),
-        chain_id: 137, // Different chain!
-        signature: original.signature.clone(),
-    };
-
-    let original_hash = hash_for_signing(&original);
-    let replayed_hash = hash_for_signing(&replayed);
-
-    assert_ne!(
-        original_hash, replayed_hash,
-        "Cross-chain replay should produce different hash"
-    );
-
-    println!("✅ Cross-chain replay prevented via EIP-155 chain ID");
-}
-
-// ═══════════════════════════════════════════════════════════════════════════════
-//                          QC-16: API GATEWAY ZERO-DAY ATTACKS
-// ═══════════════════════════════════════════════════════════════════════════════
-
-/// ZERO-DAY: JSON-RPC Parameter Injection
-///
-/// Attack: Inject malicious parameters to execute unintended operations.
-#[test]
-fn zeroday_qc16_jsonrpc_injection() {
-    // Simulate parameter validation
-    fn validate_eth_call_params(params_str: &str) -> Result<(), &'static str> {
-        // Basic JSON structure check
-        if !params_str.starts_with('[') || !params_str.ends_with(']') {
-            return Err("Params must be array");
-        }
-
-        // Defense: Check for injection characters in addresses
-        if params_str.contains("'; DROP")
-            || params_str.contains("../")
-            || params_str.contains('`')
-            || params_str.contains("GGGG")
-        // Invalid hex
-        {
-            return Err("Suspicious characters detected");
-        }
-
-        // Defense: Validate 'to' address format if present
-        if let Some(to_start) = params_str.find("\"to\":\"") {
-            let to_value_start = to_start + 6;
-            if let Some(to_end) = params_str[to_value_start..].find('"') {
-                let to_value = &params_str[to_value_start..to_value_start + to_end];
-
-                // Must start with 0x and be 42 chars (or be special address)
-                if !to_value.starts_with("0x") {
-                    return Err("Address must start with 0x");
-                }
-
-                // Check for non-hex characters after 0x
-                let hex_part = &to_value[2..];
-                if !hex_part.chars().all(|c| c.is_ascii_hexdigit()) {
-                    return Err("Address contains non-hex characters");
-                }
-            }
-        }
-
-        Ok(())
-    }
-
-    // Attack vectors
-    let attacks: Vec<(&str, &str)> = vec![
-        // SQL injection in address
-        ("sql_injection", r#"[{"to":"0x'; DROP TABLE blocks;--"}]"#),
-        // Path traversal
-        ("path_traversal", r#"[{"to":"0x../../../etc/passwd"}]"#),
-        // Command injection
-        ("cmd_injection", r#"[{"to":"0x`cat /etc/passwd`"}]"#),
-        // Invalid hex in data
-        (
-            "invalid_hex",
-            r#"[{"to":"0x0000000000000000000000000000000000000001","data":"0xGGGG"}]"#,
-        ),
-        // Missing 0x prefix
-        (
-            "missing_prefix",
-            r#"[{"to":"0000000000000000000000000000000000000001"}]"#,
-        ),
-    ];
-
-    for (name, params_str) in &attacks {
-        let result = validate_eth_call_params(params_str);
-        assert!(
-            result.is_err(),
-            "VULNERABILITY: {} attack not blocked!",
-            name
-        );
-        println!("✅ {} blocked: {}", name, result.unwrap_err());
-    }
-}
-
-/// ZERO-DAY: WebSocket Subscription Hijacking
-///
-/// Attack: Guess subscription IDs to receive data intended for other clients.
-#[test]
-fn zeroday_qc16_websocket_subscription_hijack() {
-    use std::collections::HashMap;
-
-    struct SubscriptionManager {
-        subscriptions: HashMap<String, String>, // sub_id -> connection_id
-    }
-
-    impl SubscriptionManager {
-        fn new() -> Self {
-            Self {
-                subscriptions: HashMap::new(),
-            }
-        }
-
-        fn create_subscription(&mut self, connection_id: &str) -> String {
-            // Defense: Use cryptographically random subscription IDs
-            let sub_id = format!("0x{:032x}", rand_u64() as u128 | (rand_u64() as u128) << 64);
-            self.subscriptions
-                .insert(sub_id.clone(), connection_id.to_string());
-            sub_id
-        }
-
-        fn get_data(&self, sub_id: &str, requester_connection: &str) -> Result<&str, &'static str> {
-            match self.subscriptions.get(sub_id) {
-                Some(owner) if owner == requester_connection => Ok("data"),
-                Some(_) => Err("Subscription belongs to different connection"),
-                None => Err("Unknown subscription"),
-            }
-        }
-    }
-
-    let mut manager = SubscriptionManager::new();
-
-    // Legitimate subscription
-    let legit_sub = manager.create_subscription("connection_1");
-
-    // Attack: Try to access with different connection
-    let result = manager.get_data(&legit_sub, "connection_2");
-    assert!(result.is_err(), "Subscription hijack should be blocked");
-
-    // Attack: Try to guess subscription ID
-    for i in 0..1000 {
-        let guessed_id = format!("0x{:032x}", i);
-        let result = manager.get_data(&guessed_id, "attacker");
-        assert!(result.is_err(), "Guessed subscription ID should fail");
-    }
-
-    println!("✅ WebSocket subscription hijacking blocked");
-}
-
-// ═══════════════════════════════════════════════════════════════════════════════
-//                          HELPER FUNCTIONS
-// ═══════════════════════════════════════════════════════════════════════════════
-
-fn simple_hash(data: &[u8], salt: u8) -> usize {
-    let mut hash: usize = salt as usize;
-    for &byte in data {
-        hash = hash.wrapping_mul(31).wrapping_add(byte as usize);
-    }
-    hash
-}
-
-fn rand_u64() -> u64 {
-    use std::time::{SystemTime, UNIX_EPOCH};
-    let duration = SystemTime::now()
-        .duration_since(UNIX_EPOCH)
-        .unwrap_or_default();
-    duration.as_nanos() as u64 ^ (duration.subsec_nanos() as u64).rotate_left(32)
-}
-
-// ═══════════════════════════════════════════════════════════════════════════════
-//                          SUMMARY TEST
-// ═══════════════════════════════════════════════════════════════════════════════
-
-/// Meta-test: Verify all zero-day categories are covered
-#[test]
-fn zeroday_coverage_summary() {
-    println!("\n");
-    println!("═══════════════════════════════════════════════════════════════════");
-    println!("                    ZERO-DAY ATTACK COVERAGE SUMMARY               ");
-    println!("═══════════════════════════════════════════════════════════════════");
-    println!();
-    println!("  qc-01 Peer Discovery:");
-    println!("    ✅ Eclipse via Kademlia poisoning");
-    println!("    ✅ Sybil with colluding node IDs");
-    println!("    ✅ Node ID theft/impersonation");
-    println!();
-    println!("  qc-02 Block Storage:");
-    println!("    ✅ Block hash collision attack");
-    println!("    ✅ Bit flip corruption detection");
-    println!("    ✅ Assembly race condition (TOCTOU)");
-    println!();
-    println!("  qc-03 Transaction Indexing:");
-    println!("    ✅ Merkle proof length extension");
-    println!("    ✅ Second preimage attack");
-    println!();
-    println!("  qc-04 State Management:");
-    println!("    ✅ Patricia trie key collision");
-    println!("    ✅ State proof malformed path");
-    println!();
-    println!("  qc-05 Block Propagation:");
-    println!("    ✅ Bandwidth amplification");
-    println!("    ✅ Gossip protocol poisoning");
-    println!();
-    println!("  qc-06 Mempool:");
-    println!("    ✅ Transaction hash malleability");
-    println!("    ✅ Mempool eviction attack");
-    println!();
-    println!("  qc-07 Bloom Filters:");
-    println!("    ✅ Filter saturation attack");
-    println!();
-    println!("  qc-08 Consensus:");
-    println!("    ✅ Long-range attack");
-    println!("    ✅ VRF grinding attack");
-    println!();
-    println!("  qc-09 Finality:");
-    println!("    ✅ Finality reversion attack");
-    println!("    ✅ Checkpoint manipulation");
-    println!();
-    println!("  qc-10 Signature Verification:");
-    println!("    ✅ Invalid curve point attack");
-    println!("    ✅ Cross-chain signature replay");
-    println!();
-    println!("  qc-16 API Gateway:");
-    println!("    ✅ JSON-RPC parameter injection");
-    println!("    ✅ WebSocket subscription hijacking");
-    println!();
-    println!("═══════════════════════════════════════════════════════════════════");
-    println!("  TOTAL: 18 Zero-Day Attack Categories Tested");
-    println!("═══════════════════════════════════════════════════════════════════");
-}
