+++ tests/src/exploits/architectural/cross_cutting/zero_day/consensus.rs
@@ -1,0 +1,206 @@
+use sha3::{Digest, Keccak256};
+use std::collections::HashSet;
+
+// ═══════════════════════════════════════════════════════════════════════════════
+//                          QC-08 & QC-09: CONSENSUS & FINALITY ATTACKS
+// ═══════════════════════════════════════════════════════════════════════════════
+
+#[test]
+fn zeroday_qc08_long_range_attack() {
+    // Defense: Weak subjectivity checkpoints
+    struct ChainValidator {
+        checkpoint_height: u64,
+        #[allow(dead_code)]
+        checkpoint_hash: [u8; 32],
+    }
+
+    impl ChainValidator {
+        fn validate_chain(
+            &self,
+            _chain_tip_height: u64,
+            genesis_height: u64,
+        ) -> Result<(), &'static str> {
+            // Defense: Reject chains that fork before the checkpoint
+            if genesis_height < self.checkpoint_height {
+                return Err("Chain forks before weak subjectivity checkpoint");
+            }
+            Ok(())
+        }
+    }
+
+    let validator = ChainValidator {
+        checkpoint_height: 10000,
+        checkpoint_hash: [0xAAu8; 32],
+    };
+
+    // Attack: Present chain starting from height 0
+    let result = validator.validate_chain(15000, 0);
+    assert!(result.is_err(), "Long-range attack should be rejected");
+
+    // Legitimate chain starting after checkpoint
+    let result = validator.validate_chain(15000, 10001);
+    assert!(result.is_ok(), "Valid chain should be accepted");
+
+    println!("✅ Long-range attack blocked via weak subjectivity checkpoint");
+}
+
+#[test]
+fn zeroday_qc08_vrf_grinding() {
+    // Simulate VRF output
+    fn vrf_output(secret: &[u8], input: &[u8]) -> [u8; 32] {
+        let mut hasher = Keccak256::new();
+        hasher.update(secret);
+        hasher.update(input);
+        let result = hasher.finalize();
+        let mut h = [0u8; 32];
+        h.copy_from_slice(&result);
+        h
+    }
+
+    // Attacker tries to grind for favorable output
+    let attacker_secret = b"attacker_key";
+    let slot = 1000u64;
+
+    // Defense: Include unpredictable entropy in VRF input
+    let prev_block_hash = [0xABu8; 32]; // From previous block
+
+    let mut favorable_count = 0;
+    let threshold = [0x10u8; 32]; // ~6% chance to be "leader"
+
+    for nonce in 0..1000u64 {
+        let mut input = Vec::new();
+        input.extend(&slot.to_le_bytes());
+        input.extend(&prev_block_hash);
+        input.extend(&nonce.to_le_bytes()); // Grinding parameter
+
+        let output = vrf_output(attacker_secret, &input);
+
+        if output < threshold {
+            favorable_count += 1;
+        }
+    }
+
+    // With proper VRF, grinding shouldn't significantly improve odds
+    // Expected: ~60 favorable (6% of 1000)
+    let expected = 60;
+    let tolerance = 30; // Statistical variance
+
+    assert!(
+        (favorable_count as i64 - expected as i64).abs() < tolerance as i64,
+        "Grinding attack may be effective: {} favorable vs {} expected",
+        favorable_count,
+        expected
+    );
+
+    println!(
+        "✅ VRF grinding ineffective: {} favorable outputs (expected ~{})",
+        favorable_count, expected
+    );
+}
+
+#[test]
+fn zeroday_qc09_finality_reversion() {
+    struct FinalityGadget {
+        finalized_blocks: HashSet<[u8; 32]>,
+        #[allow(dead_code)]
+        finality_threshold: f64, // 2/3 of validators
+    }
+
+    impl FinalityGadget {
+        fn is_finalized(&self, block_hash: &[u8; 32]) -> bool {
+            self.finalized_blocks.contains(block_hash)
+        }
+
+        fn try_revert(&self, block_hash: &[u8; 32], revert_votes: f64) -> Result<(), &'static str> {
+            if self.is_finalized(block_hash) {
+                // Defense: Once finalized, cannot be reverted
+                // Would require >1/3 validators to be slashed
+                if revert_votes > 0.33 {
+                    return Err("Reversion would require slashing >1/3 validators");
+                }
+                return Err("Block is finalized - cannot revert");
+            }
+            Ok(())
+        }
+    }
+
+    let mut gadget = FinalityGadget {
+        finalized_blocks: HashSet::new(),
+        finality_threshold: 0.67,
+    };
+
+    let block = [0xAAu8; 32];
+    gadget.finalized_blocks.insert(block);
+
+    // Attack: Try to revert with 40% of validators
+    let result = gadget.try_revert(&block, 0.40);
+    assert!(result.is_err(), "Finality reversion should be blocked");
+
+    println!("✅ Finality reversion attack blocked - requires slashing >1/3 stake");
+}
+
+#[test]
+fn zeroday_qc09_checkpoint_manipulation() {
+    struct Checkpoint {
+        #[allow(dead_code)]
+        epoch: u64,
+        #[allow(dead_code)]
+        block_hash: [u8; 32],
+        aggregate_signature: Vec<u8>,
+        participating_validators: Vec<[u8; 32]>,
+    }
+
+    fn verify_checkpoint(
+        checkpoint: &Checkpoint,
+        validator_set: &HashSet<[u8; 32]>,
+        threshold: f64,
+    ) -> Result<(), &'static str> {
+        // Defense 1: Verify all signers are validators
+        for validator in &checkpoint.participating_validators {
+            if !validator_set.contains(validator) {
+                return Err("Unknown validator in checkpoint");
+            }
+        }
+
+        // Defense 2: Verify threshold participation
+        let participation =
+            checkpoint.participating_validators.len() as f64 / validator_set.len() as f64;
+        if participation < threshold {
+            return Err("Insufficient participation");
+        }
+
+        // Defense 3: Verify aggregate signature (simplified)
+        if checkpoint.aggregate_signature.is_empty() {
+            return Err("Missing aggregate signature");
+        }
+
+        Ok(())
+    }
+
+    // Create validator set
+    let validator_set: HashSet<[u8; 32]> = (0..100u8).map(|i| [i; 32]).collect();
+
+    // Attack 1: Fake checkpoint with unknown validators
+    let fake_checkpoint = Checkpoint {
+        epoch: 100,
+        block_hash: [0xEEu8; 32],
+        aggregate_signature: vec![1; 96],
+        participating_validators: (200u8..=255).chain(0u8..15).map(|i| [i; 32]).collect(), // Unknown!
+    };
+
+    let result = verify_checkpoint(&fake_checkpoint, &validator_set, 0.67);
+    assert!(result.is_err(), "Fake validators should be rejected");
+
+    // Attack 2: Checkpoint with insufficient participation
+    let weak_checkpoint = Checkpoint {
+        epoch: 100,
+        block_hash: [0xAAu8; 32],
+        aggregate_signature: vec![1; 96],
+        participating_validators: (0..30u8).map(|i| [i; 32]).collect(), // Only 30%
+    };
+
+    let result = verify_checkpoint(&weak_checkpoint, &validator_set, 0.67);
+    assert!(result.is_err(), "Weak checkpoint should be rejected");
+
+    println!("✅ Checkpoint manipulation attacks blocked");
+}
