+++ tests/src/exploits/modern/qc_02/block_storage.rs
@@ -45,7 +45,7 @@ use qc_02_block_storage::ports::outbound::{
     BincodeBlockSerializer, DefaultChecksumProvider, InMemoryKVStore, MockFileSystemAdapter,
     SystemTimeSource,
 };
-use qc_02_block_storage::service::subsystem_ids;
+use qc_02_block_storage::service::{subsystem_ids, BlockStorageDependencies};
 use qc_02_block_storage::{
     AssemblyConfig, BlockAssemblyBuffer, BlockStorageApi, BlockStorageService, StorageConfig,
     StorageError,
@@ -64,14 +64,14 @@ fn make_test_service() -> BlockStorageService<
     SystemTimeSource,
     BincodeBlockSerializer,
 > {
-    BlockStorageService::new(
-        InMemoryKVStore::new(),
-        MockFileSystemAdapter::new(50), // 50% disk space
-        DefaultChecksumProvider,
-        SystemTimeSource,
-        BincodeBlockSerializer,
-        StorageConfig::default(),
-    )
+    let deps = BlockStorageDependencies {
+        kv_store: InMemoryKVStore::new(),
+        fs_adapter: MockFileSystemAdapter::new(50),
+        checksum: DefaultChecksumProvider,
+        time_source: SystemTimeSource,
+        serializer: BincodeBlockSerializer,
+    };
+    BlockStorageService::new(deps, StorageConfig::default())
 }
 
 /// Create a test service with specific disk space
@@ -84,14 +84,14 @@ fn make_service_with_disk_space(
     SystemTimeSource,
     BincodeBlockSerializer,
 > {
-    BlockStorageService::new(
-        InMemoryKVStore::new(),
-        MockFileSystemAdapter::new(percent),
-        DefaultChecksumProvider,
-        SystemTimeSource,
-        BincodeBlockSerializer,
-        StorageConfig::default(),
-    )
+    let deps = BlockStorageDependencies {
+        kv_store: InMemoryKVStore::new(),
+        fs_adapter: MockFileSystemAdapter::new(percent),
+        checksum: DefaultChecksumProvider,
+        time_source: SystemTimeSource,
+        serializer: BincodeBlockSerializer,
+    };
+    BlockStorageService::new(deps, StorageConfig::default())
 }
 
 /// Create a test service with custom assembly config
@@ -108,14 +108,14 @@ fn make_service_with_assembly_config(
         assembly_config: config,
         ..Default::default()
     };
-    BlockStorageService::new(
-        InMemoryKVStore::new(),
-        MockFileSystemAdapter::new(50),
-        DefaultChecksumProvider,
-        SystemTimeSource,
-        BincodeBlockSerializer,
-        storage_config,
-    )
+    let deps = BlockStorageDependencies {
+        kv_store: InMemoryKVStore::new(),
+        fs_adapter: MockFileSystemAdapter::new(50),
+        checksum: DefaultChecksumProvider,
+        time_source: SystemTimeSource,
+        serializer: BincodeBlockSerializer,
+    };
+    BlockStorageService::new(deps, storage_config)
 }
 
 /// Create a validated block for testing
